requirements.txt
discord.py==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
aiohttp==3.9.1
PyYAML==6.0.1 